move some functions into safe.py
