+++ pyanalyze/analysis_lib.py
@@ -5,7 +5,7 @@ Commonly useful components for static analysis tools.
 """
 import ast
 import os
-from typing import List, Callable, Optional, Set, TypeVar, Container, Type, Iterable
+from typing import List, Callable, Optional, Set, TypeVar, Type, Iterable
 from typing_extensions import Annotated
 
 from .extensions import ParameterTypeGuard
@@ -84,24 +84,6 @@ def get_line_range_for_node(node: ast.AST, lines: List[str]) -> List[int]:
     ):
         last_lineno += 1
     return list(range(first_lineno, last_lineno))
-
-
-def is_iterable(obj: object) -> bool:
-    """Returns whether a Python object is iterable."""
-    typ = type(obj)
-    if hasattr(typ, "__iter__"):
-        return True
-    return hasattr(typ, "__getitem__") and hasattr(typ, "__len__")
-
-
-def safe_in(item: T, collection: Container[T]) -> bool:
-    """Safely checks whethe item is in collection. Defaults to returning false."""
-    # Workaround against mock objects sometimes throwing ValueError if you compare them,
-    # and against objects throwing other kinds of errors if you use in.
-    try:
-        return item in collection
-    except Exception:
-        return False
 
 
 def all_of_type(
