+++ pyanalyze/test_analysis_lib.py
@@ -1,8 +1,7 @@
 import ast
-from typing import Iterable
 from qcore.asserts import assert_eq
 
-from .analysis_lib import get_indentation, get_line_range_for_node, is_iterable
+from .analysis_lib import get_indentation, get_line_range_for_node
 
 
 def test_get_indentation() -> None:
@@ -44,46 +43,3 @@ def test_get_line_range_for_node() -> None:
     assert_eq([6, 7, 8, 9, 10], get_line_range_for_node(tree.body[2], lines))
     assert_eq([13, 14], get_line_range_for_node(tree.body[3], lines))
     assert_eq([16, 17, 18, 19, 20, 21], get_line_range_for_node(tree.body[4], lines))
-
-
-def test_is_iterable() -> None:
-    def gen() -> Iterable[None]:
-        yield
-
-    class NoSpecialMethods(object):
-        pass
-
-    class HasIter(object):
-        def __iter__(self) -> Iterable[int]:
-            yield 1
-
-    class HasGetItemAndLen(object):
-        def __getitem__(self, i: int) -> int:
-            return i ** 2
-
-        def __len__(self) -> int:
-            return 1 << 15
-
-    class HasGetItem(object):
-        def __getitem__(self, i: int) -> int:
-            raise KeyError("tricked you, I am not iterable")
-
-    class HasLen(object):
-        def __len__(self) -> int:
-            return -1
-
-    assert is_iterable("")
-    assert is_iterable([])
-    assert is_iterable(range(1))
-    assert is_iterable(gen())
-    assert is_iterable({})
-    assert is_iterable({}.keys())
-    assert not is_iterable(42)
-    assert not is_iterable(None)
-    assert not is_iterable(False)
-    assert not is_iterable(str)
-    assert not is_iterable(NoSpecialMethods())
-    assert is_iterable(HasIter())
-    assert is_iterable(HasGetItemAndLen())
-    assert not is_iterable(HasGetItem())
-    assert not is_iterable(HasLen())
